Require username, cleaning
