+++ example/cypress/integration/api_test_spec.js
@@ -7,6 +7,7 @@ describe("API Endpoint tests #1", () => {
   const inputOrganization = "test"
   const inputCourse = "python-test"
   const inputExercise = "osa01-01_hymio"
+  const inputUsername = "test"
   const inputToken = "49a491a3fc7"
 
   beforeEach(() => {
@@ -16,6 +17,7 @@ describe("API Endpoint tests #1", () => {
     )
     window.localStorage.setItem("course", JSON.stringify(inputCourse))
     window.localStorage.setItem("exercise", JSON.stringify(inputExercise))
+    window.localStorage.setItem("username", JSON.stringify(inputUsername))
     window.localStorage.setItem("token", JSON.stringify(inputToken))
 
     cy.intercept(
@@ -135,6 +137,7 @@ describe("API Endpoint tests #2", () => {
   const inputOrganization = "test"
   const inputCourse = "python-test"
   const inputExercise = "osa01-01_hymio"
+  const inputUsername = "test"
   const inputToken = "49a491a3fc7"
 
   beforeEach(() => {
@@ -144,6 +147,7 @@ describe("API Endpoint tests #2", () => {
     )
     window.localStorage.setItem("course", JSON.stringify(inputCourse))
     window.localStorage.setItem("exercise", JSON.stringify(inputExercise))
+    window.localStorage.setItem("username", JSON.stringify(inputUsername))
     window.localStorage.setItem("token", JSON.stringify(inputToken))
 
     cy.intercept(
