+++ example/cypress/integration/home_page_spec.js
@@ -6,6 +6,7 @@ describe("The Playground", () => {
   const inputOrganization = "test"
   const inputCourse = "python-test"
   const inputExercise = "osa01-01_hymio"
+  const inputUsername = "test"
   const inputToken = "3213hddf"
 
   beforeEach(() => {
@@ -15,6 +16,7 @@ describe("The Playground", () => {
     )
     window.localStorage.setItem("course", JSON.stringify(inputCourse))
     window.localStorage.setItem("exercise", JSON.stringify(inputExercise))
+    window.localStorage.setItem("username", JSON.stringify(inputUsername))
     window.localStorage.setItem("token", JSON.stringify(inputToken))
 
     cy.intercept(
